% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.listofmerMod.R
\name{anova.listofmerMod}
\alias{anova.listofmerMod}
\title{Anova method for listofmerMod object}
\usage{
\method{anova}{listofmerMod}(object, ..., type = 3, ddf = "Satterthwaite")
}
\arguments{
\item{object}{a listofmerMod object.}

\item{...}{orther arguments.}

\item{type}{numerical indicating the type of SS. Default is 3.}

\item{ddf}{character indicating the computation of the degree of freedom. Default is "Satterthwaite".}
}
\description{
Apply the anova function from lmerTest at each model of the list.
}
\details{
see anova from lmerTest for more informations.
}
