% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getME.listofmerMod.R
\name{getME.listofmerMod}
\alias{getME.listofmerMod}
\title{Extract component from list of Mermod}
\usage{
\method{getME}{listofmerMod}(object, name, ...)
}
\arguments{
\item{object}{a listofMermod object.}

\item{name}{the name of the component.}

\item{...}{additional argument like SIMPLIFY}
}
\description{
Method to extract component of all the the models produce by lmersignal
}
\details{
see \link{getME} for more info about the component available.
}
