% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmersignal.R
\name{lmersignal}
\alias{lmersignal}
\title{Linear Mixed Models for signal}
\usage{
lmersignal(formula, data = NULL, REML = TRUE, control = lmerControl(),
  start = NULL, verbose = 0L, subset, weights, na.action, offset,
  contrasts = NULL, devFunOnly = FALSE, ...)
}
\arguments{
\item{formula}{a \code{lme4} formula where the response is a matrix.}

\item{data}{a data frame. See \code{\link{lmer}} for more details.}

\item{REML}{a logical that indicate which criterion to optimize. See \code{\link{lmer}} for more details.}

\item{control}{Some parameters. See \link{lmerControl} or \code{\link{lmer}} for more details.}

\item{start}{starting values for the parameters.}

\item{verbose}{See \code{\link{lmer}} for more details.}

\item{subset}{an expression to selecte a subset of the data. See \code{\link{lmer}} for more details.}

\item{weights}{an optional vector of weights. See \code{\link{lmer}} for more details.}

\item{na.action}{a function that handle \code{NA}'s. See \code{\link{lmer}} for more details.}

\item{offset}{specify a priori component in the predictor. See \code{\link{lmer}} for more details.}

\item{contrasts}{a list of coFntrasts. See \code{\link{lmer}} for more details.}

\item{devFunOnly}{a logical set by default to \code{FALSE}. See \code{\link{lmer}} for more details.}

\item{...}{addition arguments. See \code{\link{lmer}} for more details.}
}
\value{
A listofmerMod object which is a list of merMod object.
}
\description{
This function return a list mixed model fitted using the same design on each column of a matrix response variable. Theta parameters are assume to be smooth.
}
\seealso{
\code{\link{lmer}}.
}
