% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.listofMermod.R
\name{ranef.listofMermod}
\alias{ranef.listofMermod}
\title{Extract the BLUP}
\usage{
\method{ranef}{listofMermod}(object)
}
\arguments{
\item{object}{a listofMermod object.}
}
\value{
a list containing the the BLUP for all models.
}
\description{
Extract the BLUP of all model
}
