% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.listofmerMod.R
\name{ranef.listofmerMod}
\alias{ranef.listofmerMod}
\title{Extract the BLUP}
\usage{
\method{ranef}{listofmerMod}(object)
}
\arguments{
\item{object}{a listofmerMod object.}
}
\value{
a list containing the the BLUP for all models.
}
\description{
Extract the BLUP of all model
}
